requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
chardet==3.0.4
charset-normalizer==2.0.7
contourpy==1.2.0
cycler==0.12.1
deep-translator==1.11.4
Django==3.2.6
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.12.4
facebook-sdk==3.1.0
fonttools==4.49.0
french-color-to-hex==0.1.6
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.3.1
html5lib==1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
importlib_resources==6.1.2
kiwisolver==1.4.5
lxml==4.9.1
matplotlib==3.8.3
mysqlclient==2.0.3
numpy==1.21.2
packaging==23.2
Pillow==8.4.0
psycopg2==2.9.9
pyparsing==3.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2021.1
requests==2.26.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.2.1
sqlparse==0.4.1
typing_extensions==4.10.0
Unidecode==1.3.8
urllib3==1.26.7
webencodings==0.5.1
zipp==3.17.0
